start making moves to floating-point cost system
git-svn-id: svn+ssh://svn.abisource.com/svnroot/link-grammar/trunk@27617 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ link-grammar/build-disjuncts.h
@@ -10,6 +10,6 @@
 /* forms, with or without modification, subject to certain conditions.          */
 /*                                                                              */
 /********************************************************************************/
-void build_sentence_disjuncts(Sentence sent, int cost_cutoff);
+void build_sentence_disjuncts(Sentence sent, float cost_cutoff);
 X_node *   build_word_expressions(Sentence sent, const char *);
 Disjunct * build_disjuncts_for_dict_node(Dict_node *);
